Remove App.jsx
